+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod controller;
 pub mod discovery;
 pub mod domain;
 pub mod hdhr;
+pub mod logging;
 #[cfg(feature = "playback")]
 pub mod playback;
 pub mod settings;